third_party/blink/web_tests/virtual/unload-allowed/external/wpt/html/webappapis/scripting/events/event-handler-attributes-frameset-window-expected.txt

This is a testharness.js-based test.
Found 22 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] shadowed pagereveal (document.body)
  assert_equals: alternative body should reflect expected (function) function "() => 0" but got (undefined) undefined
[FAIL] shadowed pageswap (document.body)
  assert_equals: alternative body should reflect expected (function) function "() => 0" but got (undefined) undefined
[FAIL] not shadowed copy (document.body)
  assert_equals: window should reflect expected (object) null but got (undefined) undefined
[FAIL] not shadowed cut (document.body)
  assert_equals: window should reflect expected (object) null but got (undefined) undefined
[FAIL] not shadowed paste (document.body)
  assert_equals: window should reflect expected (object) null but got (undefined) undefined
[FAIL] shadowed pagereveal removal (document.body)
  assert_equals: body should reflect changes to itself expected (object) null but got (undefined) undefined
[FAIL] shadowed pageswap removal (document.body)
  assert_equals: body should reflect changes to itself expected (object) null but got (undefined) undefined
[FAIL] shadowed messageerror on body fires when event dispatched on window
  assert_true: expected true got undefined
[FAIL] shadowed pagereveal on body fires when event dispatched on window
  assert_true: expected true got undefined
[FAIL] shadowed pageswap on body fires when event dispatched on window
  assert_true: expected true got undefined
[FAIL] shadowed rejectionhandled on body fires when event dispatched on window
  assert_true: expected true got undefined
[FAIL] shadowed pagereveal (document.createElement("frameset"))
  assert_equals: body should reflect expected (function) function "() => 0" but got (undefined) undefined
[FAIL] shadowed pageswap (document.createElement("frameset"))
  assert_equals: body should reflect expected (function) function "() => 0" but got (undefined) undefined
[FAIL] not shadowed copy (document.createElement("frameset"))
  assert_equals: window should reflect expected (object) null but got (undefined) undefined
[FAIL] not shadowed cut (document.createElement("frameset"))
  assert_equals: window should reflect expected (object) null but got (undefined) undefined
[FAIL] not shadowed paste (document.createElement("frameset"))
  assert_equals: window should reflect expected (object) null but got (undefined) undefined
[FAIL] shadowed pagereveal removal (document.createElement("frameset"))
  assert_equals: alternative body should reflect changes to itself expected (object) null but got (undefined) undefined
[FAIL] shadowed pageswap removal (document.createElement("frameset"))
  assert_equals: alternative body should reflect changes to itself expected (object) null but got (undefined) undefined
[FAIL] shadowed pagereveal (window)
  assert_equals: body should reflect expected (function) function "() => 0" but got (undefined) undefined
[FAIL] shadowed pageswap (window)
  assert_equals: body should reflect expected (function) function "() => 0" but got (undefined) undefined
[FAIL] shadowed pagereveal removal (window)
  assert_equals: body should reflect expected (object) null but got (undefined) undefined
[FAIL] shadowed pageswap removal (window)
  assert_equals: body should reflect expected (object) null but got (undefined) undefined
Harness: the test ran to completion.
